requirements.txt
opencv-python
cmake
pyyaml
streamlit